+++ packages/helia/src/storage.ts
@@ -1,19 +1,15 @@
-import filter from 'it-filter'
+import createMortice from 'mortice'
+import type { Blocks, Pair, DeleteManyBlocksProgressEvents, DeleteBlockProgressEvents, GetBlockProgressEvents, GetManyBlocksProgressEvents, PutManyBlocksProgressEvents, PutBlockProgressEvents, GetAllBlocksProgressEvents, GetOfflineOptions } from '@helia/interface/blocks'
+import type { Pins } from '@helia/interface/pins'
+import type { AbortOptions } from '@libp2p/interface'
 import type { Blockstore } from 'interface-blockstore'
-import type { Blocks, Pair, DeleteManyBlocksProgressEvents, DeleteBlockProgressEvents, GetBlockProgressEvents, GetManyBlocksProgressEvents, PutManyBlocksProgressEvents, PutBlockProgressEvents, GetAllBlocksProgressEvents } from '@helia/interface/blocks'
-import type { Bitswap } from 'ipfs-bitswap'
-import type { CID } from 'multiformats/cid'
-import type { AbortOptions } from '@libp2p/interfaces'
 import type { AwaitIterable } from 'interface-store'
 import type { Mortice } from 'mortice'
-import createMortice from 'mortice'
-import type { Pins } from '@helia/interface/pins'
-import forEach from 'it-foreach'
-import { CustomProgressEvent, ProgressOptions } from 'progress-events'
+import type { CID } from 'multiformats/cid'
+import type { ProgressOptions } from 'progress-events'
 
 export interface BlockStorageInit {
   holdGcLock?: boolean
-  bitswap?: Bitswap
 }
 
 export interface GetOptions extends AbortOptions {
@@ -28,7 +24,6 @@ export interface GetOptions extends AbortOptions {
 export class BlockStorage implements Blocks {
   public lock: Mortice
   private readonly child: Blockstore
-  private readonly bitswap?: Bitswap
   private readonly pins: Pins
 
   /**
@@ -36,7 +31,6 @@ export class BlockStorage implements Blocks {
    */
   constructor (blockstore: Blockstore, pins: Pins, options: BlockStorageInit = {}) {
     this.child = blockstore
-    this.bitswap = options.bitswap
     this.pins = pins
     this.lock = createMortice({
       singleProcess: options.holdGcLock
@@ -54,18 +48,6 @@ export class BlockStorage implements Blocks {
     const releaseLock = await this.lock.readLock()
 
     try {
-      if (await this.child.has(cid)) {
-        options.onProgress?.(new CustomProgressEvent<CID>('blocks:put:duplicate', cid))
-        return cid
-      }
-
-      if (this.bitswap?.isStarted() === true) {
-        options.onProgress?.(new CustomProgressEvent<CID>('blocks:put:bitswap:notify', cid))
-        this.bitswap.notify(cid, block, options)
-      }
-
-      options.onProgress?.(new CustomProgressEvent<CID>('blocks:put:blockstore:put', cid))
-
       return await this.child.put(cid, block, options)
     } finally {
       releaseLock()
@@ -79,23 +61,7 @@ export class BlockStorage implements Blocks {
     const releaseLock = await this.lock.readLock()
 
     try {
-      const missingBlocks = filter(blocks, async ({ cid }) => {
-        const has = await this.child.has(cid)
-
-        if (has) {
-          options.onProgress?.(new CustomProgressEvent<CID>('blocks:put-many:duplicate', cid))
-        }
-
-        return !has
-      })
-
-      const notifyEach = forEach(missingBlocks, ({ cid, block }) => {
-        options.onProgress?.(new CustomProgressEvent<CID>('blocks:put-many:bitswap:notify', cid))
-        this.bitswap?.notify(cid, block, options)
-      })
-
-      options.onProgress?.(new CustomProgressEvent('blocks:put-many:blockstore:put-many'))
-      yield * this.child.putMany(notifyEach, options)
+      yield * this.child.putMany(blocks, options)
     } finally {
       releaseLock()
     }
@@ -104,21 +70,10 @@ export class BlockStorage implements Blocks {
   /**
    * Get a block by cid
    */
-  async get (cid: CID, options: AbortOptions & ProgressOptions<GetBlockProgressEvents> = {}): Promise<Uint8Array> {
+  async get (cid: CID, options: GetOfflineOptions & AbortOptions & ProgressOptions<GetBlockProgressEvents> = {}): Promise<Uint8Array> {
     const releaseLock = await this.lock.readLock()
 
     try {
-      if (this.bitswap?.isStarted() != null && !(await this.child.has(cid))) {
-        options.onProgress?.(new CustomProgressEvent<CID>('blocks:get:bitswap:get', cid))
-        const block = await this.bitswap.want(cid, options)
-
-        options.onProgress?.(new CustomProgressEvent<CID>('blocks:get:blockstore:put', cid))
-        await this.child.put(cid, block, options)
-
-        return block
-      }
-
-      options.onProgress?.(new CustomProgressEvent<CID>('blocks:get:blockstore:get', cid))
       return await this.child.get(cid, options)
     } finally {
       releaseLock()
@@ -128,19 +83,11 @@ export class BlockStorage implements Blocks {
   /**
    * Get multiple blocks back from an (async) iterable of cids
    */
-  async * getMany (cids: AwaitIterable<CID>, options: AbortOptions & ProgressOptions<GetManyBlocksProgressEvents> = {}): AsyncIterable<Pair> {
+  async * getMany (cids: AwaitIterable<CID>, options: GetOfflineOptions & AbortOptions & ProgressOptions<GetManyBlocksProgressEvents> = {}): AsyncIterable<Pair> {
     const releaseLock = await this.lock.readLock()
 
     try {
-      options.onProgress?.(new CustomProgressEvent('blocks:get-many:blockstore:get-many'))
-      yield * this.child.getMany(forEach(cids, async (cid) => {
-        if (this.bitswap?.isStarted() === true && !(await this.child.has(cid))) {
-          options.onProgress?.(new CustomProgressEvent<CID>('blocks:get-many:bitswap:get', cid))
-          const block = await this.bitswap.want(cid, options)
-          options.onProgress?.(new CustomProgressEvent<CID>('blocks:get-many:blockstore:put', cid))
-          await this.child.put(cid, block, options)
-        }
-      }))
+      yield * this.child.getMany(cids, options)
     } finally {
       releaseLock()
     }
@@ -157,7 +104,6 @@ export class BlockStorage implements Blocks {
         throw new Error('CID was pinned')
       }
 
-      options.onProgress?.(new CustomProgressEvent<CID>('blocks:delete:blockstore:delete', cid))
       await this.child.delete(cid, options)
     } finally {
       releaseLock()
@@ -173,8 +119,7 @@ export class BlockStorage implements Blocks {
     try {
       const storage = this
 
-      options.onProgress?.(new CustomProgressEvent('blocks:delete-many:blockstore:delete-many'))
-      yield * this.child.deleteMany((async function * () {
+      yield * this.child.deleteMany((async function * (): AsyncGenerator<CID> {
         for await (const cid of cids) {
           if (await storage.pins.isPinned(cid)) {
             throw new Error('CID was pinned')
@@ -202,7 +147,6 @@ export class BlockStorage implements Blocks {
     const releaseLock = await this.lock.readLock()
 
     try {
-      options.onProgress?.(new CustomProgressEvent('blocks:get-all:blockstore:get-many'))
       yield * this.child.getAll(options)
     } finally {
       releaseLock()