+++ packages/interface/src/index.ts
@@ -14,25 +14,23 @@
  * ```
  */
 
-import type { Libp2p } from '@libp2p/interface-libp2p'
-import type { AbortOptions } from '@libp2p/interfaces'
-import type { PeerId } from '@libp2p/interface-peer-id'
-import type { Datastore } from 'interface-datastore'
-import type { Pins } from './pins.js'
-import type { ProgressEvent, ProgressOptions } from 'progress-events'
-import type { CID } from 'multiformats/cid'
 import type { Blocks } from './blocks.js'
+import type { Pins } from './pins.js'
+import type { Libp2p, AbortOptions } from '@libp2p/interface'
+import type { Datastore } from 'interface-datastore'
+import type { CID } from 'multiformats/cid'
+import type { ProgressEvent, ProgressOptions } from 'progress-events'
 
 export type { Await, AwaitIterable } from 'interface-store'
 
 /**
  * The API presented by a Helia node.
  */
-export interface Helia {
+export interface Helia<T = Libp2p> {
   /**
    * The underlying libp2p node
    */
-  libp2p: Libp2p
+  libp2p: T
 
   /**
    * Where the blocks are stored
@@ -71,12 +69,4 @@ export type GcEvents =
 
 export interface GCOptions extends AbortOptions, ProgressOptions<GcEvents> {
 
-}
-
-export interface InfoOptions extends AbortOptions {
-  /**
-   * If passed, return information about this PeerId, defaults
-   * to the ID of the current node.
-   */
-  peerId?: PeerId
 }