+++ packages/interop/test/fixtures/create-helia.browser.ts
@@ -1,11 +1,12 @@
-import { createHelia, HeliaInit } from 'helia'
-import { createLibp2p } from 'libp2p'
-import { webSockets } from '@libp2p/websockets'
-import { all } from '@libp2p/websockets/filters'
 import { noise } from '@chainsafe/libp2p-noise'
 import { yamux } from '@chainsafe/libp2p-yamux'
+import { webSockets } from '@libp2p/websockets'
+import { all } from '@libp2p/websockets/filters'
 import { MemoryBlockstore } from 'blockstore-core'
 import { MemoryDatastore } from 'datastore-core'
+import { createHelia, type HeliaInit } from 'helia'
+import { createLibp2p } from 'libp2p'
+import { identifyService } from 'libp2p/identify'
 import type { Helia } from '@helia/interface'
 
 export async function createHeliaNode (init?: Partial<HeliaInit>): Promise<Helia> {
@@ -26,8 +27,12 @@ export async function createHeliaNode (init?: Partial<HeliaInit>): Promise<Helia
       yamux()
     ],
     datastore,
-    nat: {
-      enabled: false
+    services: {
+      identify: identifyService()
+    },
+    connectionGater: {
+      // allow dialing loopback
+      denyDialMultiaddr: () => false
     }
   })