+++ packages/interop/test/hashes.spec.ts
@@ -1,15 +1,15 @@
 /* eslint-env mocha */
 
+// @ts-expect-error type config is broken
+import { sha3512 } from '@multiformats/sha3'
 import { expect } from 'aegir/chai'
+import toBuffer from 'it-to-buffer'
+import { CID } from 'multiformats/cid'
+import * as raw from 'multiformats/codecs/raw'
 import { createHeliaNode } from './fixtures/create-helia.js'
 import { createKuboNode } from './fixtures/create-kubo.js'
 import type { Helia } from '@helia/interface'
 import type { Controller } from 'ipfsd-ctl'
-import toBuffer from 'it-to-buffer'
-import { CID } from 'multiformats/cid'
-import * as raw from 'multiformats/codecs/raw'
-// @ts-expect-error type config is broken
-import { sha3512 } from '@multiformats/sha3'
 
 describe('hashes', () => {
   let helia: Helia
@@ -24,7 +24,9 @@ describe('hashes', () => {
     kubo = await createKuboNode()
 
     // connect the two nodes
-    await helia.libp2p.peerStore.addressBook.add(kubo.peer.id, kubo.peer.addresses)
+    await helia.libp2p.peerStore.merge(kubo.peer.id, {
+      multiaddrs: kubo.peer.addresses
+    })
     await helia.libp2p.dial(kubo.peer.id)
   })